Corrected module paths Add MAC validation to Network Switch MAC address Fixed CPU not being set correctly Changed Processors back to CPU (inline with Packer config) Fixed Secure Boot parameter not being applied Added Secure Boot Template option for those using Server 2016 Fixed Vlan_id not being applied to default Network Adapter Fixed Static MAC address not being applied to additional network adapters Updated examples with new parameters Updated Readme
